backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
SQLAlchemy==2.0.44
psycopg==3.3.2
psycopg-pool==3.3.0
pgvector==0.4.1
langgraph==1.0.7
langgraph-checkpoint-postgres==3.0.3
langchain-core==1.2.7
langchain-openai==1.1.7
openai==2.14.0
sentence-transformers==5.1.2
transformers==4.57.6
numpy==2.4.1
pandas==2.3.3
python-dotenv==1.2.1
tqdm==4.67.1
typing-extensions==4.15.0
